[INLONG-5456][Dashboard] Support source/sink metadata management
