+++ inlong-dashboard/src/meta/sinks/clickhouse.tsx
@@ -26,7 +26,7 @@ import i18n from '@/i18n';
 import { ColumnsType } from 'antd/es/table';
 import EditableTable, { ColumnsItemProps } from '@/components/EditableTable';
 import { excludeObject } from '@/utils';
-import { sourceDataFields } from './SourceDataFields';
+import { sourceFields } from './common/sourceFields';
 
 // ClickHouse targetType
 const clickhouseTargetTypes = [
@@ -266,7 +266,7 @@ const getForm: GetStorageFormFieldsType = (
 
 const getFieldListColumns: GetStorageColumnsType = (dataType, currentValues) => {
   return [
-    ...sourceDataFields,
+    ...sourceFields,
     {
       title: `ClickHouse${i18n.t('components.AccessHelper.StorageMetaData.Clickhouse.FieldName')}`,
       dataIndex: 'fieldName',
@@ -342,7 +342,7 @@ const getFieldListColumns: GetStorageColumnsType = (dataType, currentValues) => 
 
 const tableColumns = getForm('col') as ColumnsType;
 
-export const StorageClickhouse = {
+export const clickhouse = {
   getForm,
   getFieldListColumns,
   tableColumns,
