+++ inlong-dashboard/src/meta/sinks/hive.tsx
@@ -29,7 +29,7 @@ import EditableTable, { ColumnsItemProps } from '@/components/EditableTable';
 // import request from '@/utils/request';
 import i18n from '@/i18n';
 import { excludeObject } from '@/utils';
-import { sourceDataFields } from './SourceDataFields';
+import { sourceFields } from './common/sourceFields';
 
 // hiveFieldTypes
 const hiveFieldTypes = [
@@ -344,7 +344,7 @@ const getForm: GetStorageFormFieldsType = (
 
 const getFieldListColumns: GetStorageColumnsType = (dataType, currentValues) => {
   return [
-    ...sourceDataFields,
+    ...sourceFields,
     {
       title: `HIVE${i18n.t('components.AccessHelper.StorageMetaData.Hive.FieldName')}`,
       dataIndex: 'fieldName',
@@ -413,7 +413,7 @@ const getFieldListColumns: GetStorageColumnsType = (dataType, currentValues) => 
 
 const tableColumns = getForm('col') as ColumnsType;
 
-export const StorageHive = {
+export const hive = {
   getForm,
   getFieldListColumns,
   tableColumns,
