+++ inlong-dashboard/src/meta/sinks/postgreSql.tsx
@@ -25,29 +25,33 @@ import { ColumnsType } from 'antd/es/table';
 import EditableTable, { ColumnsItemProps } from '@/components/EditableTable';
 import i18n from '@/i18n';
 import { excludeObject } from '@/utils';
-import { sourceDataFields } from './SourceDataFields';
+import { sourceFields } from './common/sourceFields';
 
-//  oracleFieldTypes
-const oracleFieldTypes = [
-  'BINARY_FLOAT',
-  'BINARY_DOUBLE',
+// postgreSQLFieldTypes
+const postgreSQLFieldTypes = [
   'SMALLINT',
-  'FLOAT',
+  'INT2',
+  'SMALLSERIAL',
+  'SERIAL2',
+  'INTEGER',
+  'SERIAL',
+  'BIGINT',
+  'BIGSERIAL',
+  'REAL',
   'FLOAT4',
   'FLOAT8',
   'DOUBLE',
-  'REAL',
-  'NUMBER',
   'NUMERIC',
-  'DATE',
   'DECIMAL',
   'BOOLEAN',
+  'DATE',
+  'TIME',
   'TIMESTAMP',
   'CHAR',
+  'CHARACTER',
   'VARCHAR',
-  'CLOB',
-  'RAW',
-  'BLOB',
+  'TEXT',
+  'BYTEA',
 ].map(item => ({
   label: item,
   value: item,
@@ -64,14 +68,24 @@ const getForm: GetStorageFormFieldsType = (
       name: 'jdbcUrl',
       rules: [{ required: true }],
       props: {
-        placeholder: 'jdbc:oracle:thin://127.0.0.1:1521/db_name',
+        placeholder: 'jdbc:postgresql://127.0.0.1:5432/db_name',
         disabled: isEdit && [110, 130].includes(currentValues?.status),
         style: { width: 500 },
       },
     },
     {
       type: 'input',
-      label: i18n.t('components.AccessHelper.StorageMetaData.Oracle.TableName'),
+      label: i18n.t('components.AccessHelper.StorageMetaData.PostgreSQL.DbName'),
+      name: 'dbName',
+      rules: [{ required: true }],
+      props: {
+        disabled: isEdit && [110, 130].includes(currentValues?.status),
+      },
+      _inTable: true,
+    },
+    {
+      type: 'input',
+      label: i18n.t('components.AccessHelper.StorageMetaData.PostgreSQL.TableName'),
       name: 'tableName',
       rules: [{ required: true }],
       props: {
@@ -81,7 +95,7 @@ const getForm: GetStorageFormFieldsType = (
     },
     {
       type: 'input',
-      label: i18n.t('components.AccessHelper.StorageMetaData.Oracle.PrimaryKey'),
+      label: i18n.t('components.AccessHelper.StorageMetaData.PostgreSQL.PrimaryKey'),
       name: 'primaryKey',
       rules: [{ required: true }],
       props: {
@@ -151,16 +165,16 @@ const getForm: GetStorageFormFieldsType = (
 
 const getFieldListColumns: GetStorageColumnsType = (dataType, currentValues) => {
   return [
-    ...sourceDataFields,
+    ...sourceFields,
     {
-      title: `ORACLE${i18n.t('components.AccessHelper.StorageMetaData.Oracle.FieldName')}`,
+      title: `POSTGRESQL${i18n.t('components.AccessHelper.StorageMetaData.PostgreSQL.FieldName')}`,
       dataIndex: 'fieldName',
       initialValue: '',
       rules: [
         { required: true },
         {
           pattern: /^[a-z][0-9a-z_]*$/,
-          message: i18n.t('components.AccessHelper.StorageMetaData.Oracle.FieldNameRule'),
+          message: i18n.t('components.AccessHelper.StorageMetaData.PostgreSQL.FieldNameRule'),
         },
       ],
       props: (text, record, idx, isNew) => ({
@@ -168,18 +182,18 @@ const getFieldListColumns: GetStorageColumnsType = (dataType, currentValues) => 
       }),
     },
     {
-      title: `ORACLE${i18n.t('components.AccessHelper.StorageMetaData.Oracle.FieldType')}`,
+      title: `POSTGRESQL${i18n.t('components.AccessHelper.StorageMetaData.PostgreSQL.FieldType')}`,
       dataIndex: 'fieldType',
-      initialValue: oracleFieldTypes[0].value,
+      initialValue: postgreSQLFieldTypes[0].value,
       type: 'select',
       props: (text, record, idx, isNew) => ({
-        options: oracleFieldTypes,
+        options: postgreSQLFieldTypes,
         disabled: [110, 130].includes(currentValues?.status as number) && !isNew,
       }),
       rules: [{ required: true }],
     },
     {
-      title: i18n.t('components.AccessHelper.StorageMetaData.Oracle.IsMetaField'),
+      title: i18n.t('components.AccessHelper.StorageMetaData.PostgreSQL.IsMetaField'),
       initialValue: 0,
       dataIndex: 'isMetaField',
       type: 'select',
@@ -197,7 +211,7 @@ const getFieldListColumns: GetStorageColumnsType = (dataType, currentValues) => 
       }),
     },
     {
-      title: i18n.t('components.AccessHelper.StorageMetaData.Oracle.FieldFormat'),
+      title: i18n.t('components.AccessHelper.StorageMetaData.PostgreSQL.FieldFormat'),
       dataIndex: 'fieldFormat',
       initialValue: '',
       type: 'autocomplete',
@@ -211,7 +225,7 @@ const getFieldListColumns: GetStorageColumnsType = (dataType, currentValues) => 
         ['BIGINT', 'DATE', 'TIMESTAMP'].includes(record.fieldType as string),
     },
     {
-      title: i18n.t('components.AccessHelper.StorageMetaData.Oracle.FieldDescription'),
+      title: i18n.t('components.AccessHelper.StorageMetaData.PostgreSQL.FieldDescription'),
       dataIndex: 'fieldComment',
       initialValue: '',
     },
@@ -220,7 +234,7 @@ const getFieldListColumns: GetStorageColumnsType = (dataType, currentValues) => 
 
 const tableColumns = getForm('col') as ColumnsType;
 
-export const StorageOracle = {
+export const postgreSql = {
   getForm,
   getFieldListColumns,
   tableColumns,
