+++ inlong-dashboard/src/meta/sources/index.ts
@@ -1,0 +1,55 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import type { GetStorageFormFieldsType } from '@/utils/metaData';
+import type { ColumnsType } from 'antd/es/table';
+import { autoPush } from './autoPush';
+import { binLog } from './binLog';
+import { file } from './file';
+
+export interface SourceType {
+  label: string;
+  value: string;
+  // Generate form configuration for single data
+  getForm: GetStorageFormFieldsType;
+  // Generate table display configuration
+  tableColumns: ColumnsType;
+  // Custom convert interface data to front-end data format
+  toFormValues?: (values: unknown) => unknown;
+  // Custom convert front-end data to interface data format
+  toSubmitValues?: (values: unknown) => unknown;
+}
+
+export const sources: SourceType[] = [
+  {
+    label: 'MySQL BinLog',
+    value: 'MYSQL_BINLOG',
+    ...binLog,
+  },
+  {
+    label: 'File',
+    value: 'FILE',
+    ...file,
+  },
+  {
+    label: 'Auto Push',
+    value: 'AUTO_PUSH',
+    ...autoPush,
+  },
+];
