+++ inlong-dashboard/src/pages/AccessDetail/DataStorage/index.tsx
@@ -24,7 +24,7 @@ import { defaultSize } from '@/configs/pagination';
 import { useRequest } from '@/hooks';
 import i18n from '@/i18n';
 import { DataStorageDetailModal } from '@/components/AccessHelper';
-import { Storages } from '@/components/MetaData';
+import { Storages } from '@/meta/sinks';
 import request from '@/utils/request';
 import { CommonInterface } from '../common';
 import { statusList, genStatusTag } from './status';
